Caller side calculation of tile clicked + AI test code
The position of the tile clicked is not calculated on the caller side rather than in the Board class' revealTile() method. This allows revealTile() to be used in a wider range of circumstances, not just when a tile is clicked on the screen. For example, when the computer player wished to reveal a certain tile. The revealTile() method will now use the location of the tile in the ArrayList of tiles instead of the position on the screen.