revealTile() now uses the position of the Tile in ArrayList, not location on screen
This is so that revealTile() can be used in a wider range of circumstances, not just when given the coordinates on the screen where a tile is clicked. Now the method works when given a position in the ArrayList of tiles, e.g. the first tile in the arraylist is (0,0), so 0 and 0 are passed in. To get this to work, GameController has also been changed. Now, when the tile is clicked the calculation of the position in the ArrayList is done caller side.