Now sleeps between turns and works
Also some test code entered to show when a mine is found and which tile is revealed on each loop. NOTE: currently I'm not sure why but to refresh the GUI and see the update this makes, you have to minimize and then re-open the application. repaint() / revalidate() on the JFrame didn't solve this.