Update Computer.java
Now has no magic numbers again. SLEEP_TIME is the time the thread will be put to sleep for between turns. Also, a condition is added to revealing tiles, to check whether the tile is hidden or not. If tiles are not hidden (/ already revealed) then it should not count as the computer's turn. This seems to not quite be working so there may be something wrong with the reveal algorithm regarding making the tiles not hidden when they are revealed.